moved menu definers to package
