+++ src/main/java/net/minheur/potoflux/translations/register/PotoFluxTranslations.java
@@ -206,6 +206,12 @@ public class PotoFluxTranslations extends AbstractTranslationsRegistry {
         addMenu("account", "openTab")
                 .en("Open account tab")
                 .fr("Ouvrir l'onglet 'compte'");
+        addMenu("account", "login")
+                .en("Login")
+                .fr("Connexion");
+        addMenu("account", "logout")
+                .en("Logout")
+                .fr("Déconnexion");
 
         addScreen("tabHereNotHere")
                 .en("ERROR: this tab is detected but unexisting !")
